# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {},
        "license": {
            "name": "Apache 2.0"
        }
    },
    "paths": {
...
        "/api/v1/base/positions": {
            "get": {
                "description": "查询职位列表",
                "tags": [
                    "Position"
                ],
                "summary": "查询职位列表",
                "parameters": [
                    {
                        "type": "boolean",
                        "description": "是否仅查询count",
                        "name": "-",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "当前页",
                        "name": "current",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "是否查找所有数据",
                        "name": "find_all",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "职位ID集合(逗号分隔)",
                        "name": "ids",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "备注(模糊查询)",
                        "name": "memo",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "职位名称",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "职位的组织ID",
                        "name": "organization_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页大小",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "模糊搜索(模糊查询:职位名称、备注)",
                        "name": "queryValue",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "职位的角色",
                        "name": "role_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "按权重值排序(1:升序排序  2:降序排序)",
                        "name": "sequence_sort",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "状态(1:启用 2:禁用)",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ginx.ListResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ginx.OperateResult"
                        }
                    }
                }
            },
            "post": {
                "description": "创建职位",
                "tags": [
                    "Position"
                ],
                "summary": "创建职位",
                "parameters": [
                    {
                        "description": "创建职位参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schema.Position"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.IDResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ginx.OperateResult"
                        }
                    }
                }
            }
        },
        "/api/v1/base/positions/:id": {
            "get": {
                "description": "获取职位基础信息",
                "tags": [
                    "Position"
                ],
                "summary": "获取职位基础信息",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "职位ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schema.Position"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ginx.OperateResult"
                        }
                    }
                }
            },
            "put": {
                "description": "更新职位基础信息",
                "tags": [
                    "Position"
                ],
                "summary": "更新职位基础信息",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "职位ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新职位基础信息参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schema.Position"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ginx.OperateResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ginx.OperateResult"
                        }
                    }
                }
            },
            "delete": {
                "description": "删除职位",
                "tags": [
                    "Position"
                ],
                "summary": "删除职位",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "职位ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ginx.OperateResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ginx.OperateResult"
                        }
                    }
                }
            }
        },
        "/api/v1/base/positions/user-join/:id": {
            "put": {
                "description": "用户新增职位",
                "tags": [
                    "Position"
                ],
                "summary": "用户新增职位",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "职位ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "用户ID集合",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ginx.OperateResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ginx.OperateResult"
                        }
                    }
                }
            }
        },
        "/api/v1/base/positions/user-remove/:id": {
            "put": {
                "description": "用户移除职位",
                "tags": [
                    "Position"
                ],
                "summary": "用户移除职位",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "职位ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "用户ID集合",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ginx.OperateResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ginx.OperateResult"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "common.IDResult": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "0"
                }
            }
        },
        "common.PaginationResult": {
            "type": "object",
            "properties": {
                "current": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "ginx.ListResult": {
            "type": "object",
            "properties": {
                "list": {},
                "pagination": {
                    "$ref": "#/definitions/common.PaginationResult"
                }
            }
        },
        "ginx.OperateResult": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "错误码",
                    "type": "integer"
                },
                "message": {
                    "description": "错误信息",
                    "type": "string"
                }
            }
        },
...
        "schema.Position": {
            "type": "object",
            "required": [
                "name",
                "organization_id"
            ],
            "properties": {
                "created_at": {
                    "description": "创建时间",
                    "type": "string"
                },
                "creator": {
                    "description": "创建者",
                    "type": "integer"
                },
                "id": {
                    "description": "唯一标识",
                    "type": "integer"
                },
                "memo": {
                    "description": "备注",
                    "type": "string"
                },
                "name": {
                    "description": "职位名称",
                    "type": "string"
                },
                "organization_id": {
                    "description": "职位的组织ID",
                    "type": "integer"
                },
                "role_id": {
                    "description": "职位的角色",
                    "type": "integer"
                },
                "sequence": {
                    "description": "排序值",
                    "type": "integer"
                },
                "status": {
                    "description": "状态(1:启用 2:禁用)",
                    "type": "integer"
                },
                "updated_at": {
                    "description": "更新时间",
                    "type": "string"
                }
            }
        },
...
    }
}